perf: remove floats, optimize int types
Remove float division that is not necessary and a performance
bottleneck. Also reworked the code to use the optimal unsigned integer
sizes and the various stages of encoding and decoding.

Replaced some manual bit twiddling with stdlib functions for big
endian-based encoding/decoding.

Benchmark improvements results for this patch:

 * encode long string: -11% mean
 * encode long string from buffer: -6% mean
 * decode long string: -12% mean

No sensible outliers in benchmark results.