feat: performance improvements (#2)
* perf: remove floats, optimize int types

Remove float division that is not necessary and a performance
bottleneck. Also reworked the code to use the optimal unsigned integer
sizes and the various stages of encoding and decoding.

Replaced some manual bit twiddling with stdlib functions for big
endian-based encoding/decoding.

Benchmark improvements results for this patch:

 * encode long string: -11% mean
 * encode long string from buffer: -6% mean
 * decode long string: -12% mean

No sensible outliers in benchmark results.

* perf: preallocate decode buffer

Instead of reallocating the output decode buffer on the fly, we can
preallocate it with a sensible size. This will avoid multiple calls to
realloc when decoding long strings.

Benchmark improvements results for this patch:

* decode long string: -26% mean

No sensible outliers in benchmark results.